chat-service/jwt_utils.py
import jwt
from config import JWT_SECRET

def verify_token(token):
    try:
        payload = jwt.decode(token, JWT_SECRET, algorithms=["HS256"])
        return payload
    except Exception:
        return None
